OMXNodeInstance: Fix use-after-free in freeNode()
Commit "9745c8f66 OMXNodeInstance: use a lock around OMX::freeNode"
(Change-Id: I7720dd900bfa252f8675e0c56191adbf52aa957e) introduced a
use-after-free of mLock.

freeNode() deletes "this" right before it returns. When _l (a scoped Autolock)
goes out of scope _l tries to unlock mLock. mLock was already freed because it
is a member of "this." The unlock action will write to mLock, and corrupts the
heap.

This fixes the YouTube app, which was occasionally aborting with "heap
corruption" messages.

This use-after-free was detected by installing debug (with symbols)
versions of libc.so , libstagefright.so , and libstagefright_omx.so and running
Valgrind's memcheck on an ExoPlayer demo app (eu.mindtv.exoplayer.demo).

Memcheck outputs the following when a video starts to load:
Invalid write of size 4
   at 0x4826980: pthread_mutex_unlock_impl (bionic_atomic_arm.h:68)
   by 0x220A925F: android::Mutex::Autolock::~Autolock() (Mutex.h:115)
   by 0x220AD783: android::OMXNodeInstance::freeNode(android::OMXMaster*) (OMXNodeInstance.cpp:288)
   by 0x220ABB5F: android::OMX::freeNode(void*) (OMX.cpp:278)
   by 0x21FFB8E1: android::MuxOMX::freeNode(void*) (OMXClient.cpp:246)
   by 0x220002C9: android::QueryCodec(android::sp<android::IOMX> const&, char const*, char const*, bool, android::CodecCapabilities*) (OMXCodec.cpp:6363)
   by 0x21FF75C7: android::MediaCodecList::getCodecCapabilities(unsigned int, char const*, android::Vector<android::MediaCodecList::ProfileLevel>*, android::Vector<unsigned int>*, unsigned int*) const (MediaCodecList.cpp:543)
   by 0x21F55113: ??? (in /system/lib/libmedia_jni.so)
 Address 0x2299d1b0 is 0 bytes inside a block of size 80 free'd
   at 0x48972B4: operator delete[](void*) (in /system/lib/valgrind/vgpreload_memcheck-arm-linux.so)
   by 0x220AD775: android::OMXNodeInstance::freeNode(android::OMXMaster*) (OMXNodeInstance.cpp:286)
   by 0x220ABB5F: android::OMX::freeNode(void*) (OMX.cpp:278)
   by 0x21FFB8E1: android::MuxOMX::freeNode(void*) (OMXClient.cpp:246)
   by 0x220002C9: android::QueryCodec(android::sp<android::IOMX> const&, char const*, char const*, bool, android::CodecCapabilities*) (OMXCodec.cpp:6363)
   by 0x21FF75C7: android::MediaCodecList::getCodecCapabilities(unsigned int, char const*, android::Vector<android::MediaCodecList::ProfileLevel>*, android::Vector<unsigned int>*, unsigned int*) const (MediaCodecList.cpp:543)
   by 0x21F55113: ??? (in /system/lib/libmedia_jni.so)